feat(gui): Add Azure as a provider
