+++ docs/docs/reference/Model Providers/azure.md
@@ -1,0 +1,21 @@
+# Azure OpenAI
+
+If you'd like to use OpenAI models but are concerned about privacy, you can use the Azure OpenAI service, which is GDPR and HIPAA compliant.
+
+:::info[Getting access]
+You need to apply for access to the Azure OpenAI service. Response times are typically within a few days.
+
+**[Click here to apply for access to the Azure OpenAI service](https://azure.microsoft.com/en-us/products/ai-services/openai-service)**
+:::
+
+## Configuration
+
+```json title="~/.continue/config.json"
+"models": [{
+    "title": "Azure OpenAI",
+    "provider": "azure",
+    "model": "gpt-4",
+     "apiBase": "https://my-azure-openai-instance.openai.azure.com/",
+    "apiKey": "<MY_API_KEY>"
+}]
+```
