+++ docs/docs/reference/Model Providers/openai.md
@@ -2,26 +2,7 @@
 
 The OpenAI class can be used to access OpenAI models like GPT-4, GPT-4 Turbo, and GPT-3.5 Turbo.
 
-### Azure OpenAI Service
-
-If you'd like to use OpenAI models but are concerned about privacy, you can use the Azure OpenAI service, which is GDPR and HIPAA compliant. After applying for access [here](https://azure.microsoft.com/en-us/products/ai-services/openai-service), you will typically hear back within only a few days. Once you have access, set up a model in `config.json` like so:
-
-```json
-"models": [{
-    "title": "Azure OpenAI",
-    "provider": "openai",
-    "model": "gpt-4",
-    "apiBase": "https://my-azure-openai-instance.openai.azure.com/",
-    "engine": "my-azure-openai-deployment",
-    "apiVersion": "2023-07-01-preview",
-    "apiType": "azure",
-    "apiKey": "<MY_API_KEY>"
-}]
-```
-
-The easiest way to find this information is from the chat playground in the Azure OpenAI portal. Under the "Chat Session" section, click "View Code" to see each of these parameters.
-
-### OpenAI compatible servers / APIs
+## OpenAI compatible servers / APIs
 
 OpenAI compatible servers
