+++ gui/src/pages/AddNewModel/configs/completionParamsInputs.ts
@@ -1,4 +1,4 @@
-import { InputDescriptor } from "./modelPackages";
+import { InputDescriptor } from "./providers";
 
 export const contextLength: InputDescriptor = {
   inputType: "number",
