+++ gui/src/pages/AddNewModel/index.ts
@@ -1,0 +1,2 @@
+export { default as AddNewModel } from "./AddNewModel";
+export { default as ConfigureProvider } from "./ConfigureProvider";
