--- gui/src/pages/modelConfig/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ModelConfig";
